+++ src/main/java/edu/uc/rphash/StreamClusterer.java
@@ -3,7 +3,7 @@ package edu.uc.rphash;
 import java.util.List;
 
 public interface StreamClusterer extends Clusterer {
-	public abstract int addVectorOnlineStep(float[] x);
+	public abstract long addVectorOnlineStep(float[] x);
 	public abstract List<float[]> getCentroidsOfflineStep();
 
 }
